Rudimentary animation working.  NEEDS OPTIMIZATION
I've created the methods required to request the new data from
the database, and then refresh the plot.  The memory ramps up
as the plot is updated however, leading me to need to make an
optimized version.  Also because the data needs to be parsed each
update, the request takes somewhere around 1s.  which may be a
barrier to a useful program.  A better solution would be to only
pull down the new data, so perhaps sending up the latest time value
and then retrieving rows since then.  Parsing is then left to the
client and speed is affected by the client machine... Also, some
self contained calls to start and stop the updates would be nice.
Something like a button for each plot with an interval box???