Can now send only recent data, y-axis numbers
The php script can now send back only the values since the requested
time e.g. ?since=12345 or a whole 2 mins of data if no params are
are passed.  Added a first and last method for arrays.  Added numbers
for the y axis to give some sense of scale.  Still need to work on
gridlines and delineations.